requirements.txt
auth==0.5.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
eventlet==0.36.0
falcon==3.1.3
Flask==3.0.2
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.28.2
packaging==24.0
pip==22.3
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
setuptools==65.5.0
urllib3==2.2.1
Werkzeug==3.0.1
